feat: break down to smaller widgets file
