+++ lib/graphql/backend/event/mutation/update_event_reward_use.graphql
@@ -1,0 +1,3 @@
+mutation UpdateEventRewardUse($input: UpdateEventRewardUseInput!) {
+     updateEventRewardUse(input: $input)
+}
